Static typing with phpstan (#13)

+++ tests/IntegrationTests/Common/IntegrationTestCase.php
@@ -12,8 +12,10 @@ use PHPUnit\Framework\TestCase;
  * @coversNothing
  */
 class IntegrationTestCase extends TestCase {
-    private $test_server_process;
-    private $pipes;
+    /** @var ?resource */
+    private mixed $test_server_process = null;
+    /** @var ?array<resource> */
+    private ?array $pipes = null;
 
     protected function setUp(): void {
         date_default_timezone_set('UTC');
@@ -25,10 +27,14 @@ class IntegrationTestCase extends TestCase {
     private function removeRecursive(string $path): void {
         if (is_dir($path)) {
             $entries = scandir($path);
-            foreach ($entries as $entry) {
-                if ($entry !== '.' && $entry !== '..') {
-                    $entry_path = realpath("{$path}/{$entry}");
-                    $this->removeRecursive($entry_path);
+            if ($entries) {
+                foreach ($entries as $entry) {
+                    if ($entry !== '.' && $entry !== '..') {
+                        $entry_path = realpath("{$path}/{$entry}");
+                        if ($entry_path) {
+                            $this->removeRecursive($entry_path);
+                        }
+                    }
                 }
             }
             rmdir($path);
@@ -39,17 +45,21 @@ class IntegrationTestCase extends TestCase {
 
     protected function tearDown(): void {
         if ($this->test_server_process !== null) {
-            stream_set_blocking($this->pipes[1], false);
-            stream_set_blocking($this->pipes[2], false);
+            if ($this->pipes) {
+                stream_set_blocking($this->pipes[1], false);
+                stream_set_blocking($this->pipes[2], false);
+            }
             echo "Stopping server...\n";
             $this->stopTestServer();
             echo "Stopped server.\n";
-            echo "\n";
-            echo "STDOUT:\n";
-            echo fread($this->pipes[1], 1024 * 1024);
-            echo "\n";
-            echo "STDERR:\n";
-            echo fread($this->pipes[2], 1024 * 1024);
+            if ($this->pipes) {
+                echo "\n";
+                echo "STDOUT:\n";
+                echo fread($this->pipes[1], 1024 * 1024);
+                echo "\n";
+                echo "STDERR:\n";
+                echo fread($this->pipes[2], 1024 * 1024);
+            }
             echo "\n\n";
             $this->test_server_process = null;
             $this->pipes = null;
@@ -57,10 +67,10 @@ class IntegrationTestCase extends TestCase {
     }
 
     protected function startTestServer(
-        $host = '127.0.0.1',
-        $port = 8080,
-        $path = null
-    ) {
+        string $host = '127.0.0.1',
+        int $port = 8080,
+        ?string $path = null
+    ): void {
         $is_server_up = $this->isServerUp("http://{$host}:{$port}/is_server_up.html");
         if ($is_server_up) {
             throw new \Exception("A server instance is already running!");
@@ -68,17 +78,21 @@ class IntegrationTestCase extends TestCase {
         if ($path === null) {
             $path = __DIR__.'/../tmp/test-server/';
             if (!is_dir($path)) {
-                mkdir($path, 0777, true);
+                mkdir($path, 0o777, true);
             }
         }
         echo "Starting server...\n";
         $php_path = system('which php');
         $descriptorspec = [['pipe', 'r'], ['pipe', 'w'], ['pipe', 'w']];
-        $this->test_server_process = proc_open(
+        $proc = proc_open(
             "{$php_path} -S {$host}:{$port} -t {$path}",
             $descriptorspec,
             $this->pipes,
         );
+        if (!$proc) {
+            throw new \Exception("Could not proc_open.");
+        }
+        $this->test_server_process = $proc;
         $is_server_up_path = "{$path}/is_server_up.html";
         file_put_contents($is_server_up_path, 'true');
         if (is_resource($this->test_server_process)) {
@@ -101,11 +115,14 @@ class IntegrationTestCase extends TestCase {
         echo "Could not start server.\n";
     }
 
-    protected function stopTestServer() {
+    protected function stopTestServer(): void {
+        if (!$this->test_server_process) {
+            return;
+        }
         proc_terminate($this->test_server_process);
         for ($i = 0; $i < 3; $i++) {
             $status = proc_get_status($this->test_server_process);
-            $is_still_running = $status['running'] ?? false;
+            $is_still_running = $status['running'];
             if (!$is_still_running) {
                 return;
             }
@@ -115,8 +132,11 @@ class IntegrationTestCase extends TestCase {
         echo "Could not stop test server.\n";
     }
 
-    protected function isServerUp($url) {
+    protected function isServerUp(string $url): bool {
         $ch = curl_init($url);
+        if (!$ch) {
+            throw new \Exception('Could not create curl handle?!?');
+        }
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
         return curl_errno($ch) === 0 && curl_exec($ch) === 'true';
     }
